+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/DefaultRulesetTest.java
@@ -9,12 +9,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import net.sourceforge.pmd.AbstractRuleSetFactoryTest;
 import net.sourceforge.pmd.RuleSet;
 import net.sourceforge.pmd.RuleSetLoader;
 
 import com.github.stefanbirkner.systemlambda.SystemLambda;
 
 class DefaultRulesetTest {
+    private static final String QUICKSTART_RULESET = "rulesets/apex/quickstart.xml";
 
     @Test
     void loadDefaultRuleset() {
@@ -25,10 +27,15 @@ class DefaultRulesetTest {
     @Test
     void loadQuickstartRuleset() throws Exception {
         String log = SystemLambda.tapSystemErr(() -> {
-            RuleSet ruleset = rulesetLoader().loadFromResource("rulesets/apex/quickstart.xml");
+            RuleSet ruleset = rulesetLoader().loadFromResource(QUICKSTART_RULESET);
             assertNotNull(ruleset);
         });
         assertTrue(log.isEmpty(), "No Logging expected");
+    }
+
+    @Test
+    void correctEncoding() throws Exception {
+        assertTrue(AbstractRuleSetFactoryTest.hasCorrectEncoding(QUICKSTART_RULESET));
     }
 
     private RuleSetLoader rulesetLoader() {